Refactor: Remove extension package files and implement file system utilities
- Deleted `package.json` and `tsconfig.json` from the extension directory.
- Introduced `files.go` for file handling, including functions for validating files, resolving modules, and fetching from GitHub.
- Added tests for file utilities in `files_test.go`.
- Created a new `lists` package with utility functions for slice operations, including `Has`, `Some`, and `None`.
- Implemented string utility functions in `strings.go` for string manipulation and formatting.
