--- compiler/cmd/resolver/reader.go
@@ -1,13 +1,0 @@
-package resolver
-
-import (
-	"os"
-)
-
-const EXT = ".fer"
-
-// Check if file exists and is a regular file
-func IsValidFile(filename string) bool {
-	fileInfo, err := os.Stat(filename)
-	return err == nil && fileInfo.Mode().IsRegular()
-}
