+++ compiler/internal/fsutils/lists/lists_test.go
@@ -1,4 +1,4 @@
-package utils
+package lists
 
 import (
 	"testing"
