+++ compiler/internal/semantic/prelude.go
@@ -1,1 +1,1 @@
-package semantic 
+package semantic
