+++ compiler/internal/semantic/resolver.go
@@ -1,1 +1,1 @@
-package semantic 
+package semantic
