+++ compiler/internal/semantic/resolver_test.go
@@ -1,1 +1,1 @@
-package semantic 
+package semantic
