+++ compiler/internal/semantic/symbolTable.go
@@ -1,1 +1,1 @@
-package semantic 
+package semantic
