+++ compiler/internal/semantic/symbols.go
@@ -1,1 +1,1 @@
-package semantic 
+package semantic
