--- compiler/internal/serializer/serialize.go
@@ -1,1 +1,0 @@
-package serializer
