--- extension/.gitattributes
@@ -1,2 +1,0 @@
-# Set default behavior to automatically normalize line endings.
-* text=auto
