--- extension/.gitignore
@@ -1,5 +1,0 @@
-node_modules
-out
-bin
-*.vsix
-.env
