--- extension/CHANGELOG.md
@@ -1,7 +1,0 @@
-# Change Log
-
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
-## [Released]
-
-- Initial release v.0.0.1
