--- extension/README.md
@@ -1,1 +1,0 @@
-# Syntax highlighter for ferret
