--- extension/tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "es2020",
-        "outDir": "out",
-        "lib": ["es2020"],
-        "sourceMap": true,
-        "rootDir": "src",
-        "strict": true,
-        "esModuleInterop": true,
-        "skipLibCheck": true
-    },
-    "include": ["src"],
-    "exclude": ["node_modules"]
-}
